backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==4.2.11
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.0
djoser==2.2.2
django-filter
gunicorn==22.0.0
idna==3.10
oauthlib==3.2.2
packaging==24.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
urllib3==2.4.0
django-cors-headers==4.0.0